Fix #19: Don't use innerHTML
The encrypted text was inserted as innerHTML, which allows HTML content, but breaks if that is just plaintext with '<' in it.

If it is intended, that you can encrypt HTML tags, which then are inserted, don't merge this pull request. But if is is only intended for encrypted text, this would be the correct way.